+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 10.0.0
 ### New Feature
 * Add parameter query `includeMasterTerminal` and return in TerminalApi-`getTerminal`、`getTerminalBySn` API
-* Add TerminalApkParameter API `findApkParameterPidList` and return pidList
+* Add AppApi API `searchApkParameterPidList` and return pidList
 
 ### Dependency Change
 * Update guava from 31.1-jre to 33.4.8-jre